requirements.txt
flask==2.3.3
python-dotenv==1.0.0
openai==1.6.1
flask-limiter==3.5.0
redis==5.0.1
httpx>=0.24.1
